Fix PARSER_CNT evaluation on macOS
On macOS, sed does not treat '\n' as one would expect, producing a
single line of output. Using a literal newline seems to be a portable
solution for this problem (see e.g. [1]), but looks atrocious inside a
makefile. Replace commas with spaces, and count number of words instead.

In addition to that, on macOS 'wc' will right pad the number, which
causes PARSER_CNT to be set to something like "        2", breaking
"-DPARSER_CNT=$(PARSER_CNT)" expression later. Use 'tr' to remove the
extra whitespace.

[1] http://nlfiedler.github.io/2010/12/05/newlines-in-sed-on-mac.html