Fix stringop-truncation GCC warning in NMEA_PARSER_PREFIX macro
Recent versions of GCC verify that the destination of strncpy call is
sufficiently long for the string and the terminating nul byte.
Since the intention of the code is to copy the characters only,
without the nul terminator, using memcpy is more appropriate.

Also replace the numeric constant with a macro for consistency.