Fix redefinition warning for PARSER_PATH
Makefile unconditionally adds -DPARSER_PATH=... flag to the compiler
command line. This results in a preprocessor warning about redefined
variable. Fix by guarding the definition with an ifdef.

This also fixes the syntax for defining a string preprocessor macro
on the command line. Previously, -DPARSER_PATH=$(PREFIX)/lib/nmea/
would produce the following (equivalent) definition:
  #define PARSER_PATH /usr/lib/nmea/
Note the lack of quotes. This was not noticed because PARSER_PATH was
later redefined in the source code.