Enable -Werror, fix related warnings, fixes for PARSER_PATH
* Fix redefinition warning for PARSER_PATH

Makefile unconditionally adds -DPARSER_PATH=... flag to the compiler
command line. This results in a preprocessor warning about redefined
variable. Fix by guarding the definition with an ifdef.

This also fixes the syntax for defining a string preprocessor macro
on the command line. Previously, -DPARSER_PATH=$(PREFIX)/lib/nmea/
would produce the following (equivalent) definition:
  #define PARSER_PATH /usr/lib/nmea/
Note the lack of quotes. This was not noticed because PARSER_PATH was
later redefined in the source code.

* Define PARSER_PATH in CMake as it is done in Make

* Build: enable -Wall -Werror in both Make and CMake

* Fix self-assignment warning

Clang produces a warning without this fix:
    explicitly assigning value of variable of type 'char *' to itself
      [-Werror,-Wself-assign]

* Fix comparison of an array with NULL

type_word is an array member of a structure, so its address is never
a NULL. Remove the comparison.

This fixes the warning:
    comparison of array 'parsers[i].parser.type_word' equal to a null
    pointer is always false [-Werror,-Wtautological-pointer-compare]

* Fix test_parse_invalid never called

Enabling the warnings helped find that test_parse_invalid was unused.

The test did not pass initially, the commit also fixes the causes of
test failure.

* Mention NMEA_PARSER_PATH environment variable in readme file

Closes https://github.com/jacketizer/libnmea/issues/42

* ci: run tests with CMake first, set NMEA_EXPORT_PATH

Running tests with Make currently requires a system-wide installation
("sudo make install") to be done.
Run CMake tests first, so that they are not a affected by libraries
build with Make.
Also export NMEA_EXPORT_PATH so that CMake-built unit tests can find
the shared libraries.

* Fix memory leaks in test_crop_sentence_ok